Revert "Fix for a NetBSD linter warning:"
This reverts commit 1ac76ab74758a75299124bc1cf3669a0740bd5f7.

+++ pc.c
@@ -1098,7 +1098,7 @@ print_result(ULONG value)
 
       (void)fprintf(stdout, "%s", fields[i]);
 
-      line_len += (long)field_len;
+      line_len += field_len;
 
       if (fields[i+1] != NULL)
         {
